Moved TranslateRawParameters method from BaseHapiFhirResourceDao to BaseStorageDao
