+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/dao/BaseHapiFhirDao.java
@@ -133,8 +133,6 @@ public abstract class BaseHapiFhirDao<T extends IBaseResource> extends BaseStora
 	@Autowired
 	protected ISearchCoordinatorSvc mySearchCoordinatorSvc;
 	@Autowired
-	protected ISearchParamRegistry mySerarchParamRegistry;
-	@Autowired
 	protected ITermReadSvc myTerminologySvc;
 	@Autowired
 	protected IResourceHistoryTableDao myResourceHistoryTableDao;
@@ -142,8 +140,7 @@ public abstract class BaseHapiFhirDao<T extends IBaseResource> extends BaseStora
 	protected IResourceTableDao myResourceTableDao;
 	@Autowired
 	protected IResourceTagDao myResourceTagDao;
-	@Autowired
-	protected ISearchParamRegistry mySearchParamRegistry;
+
 	@Autowired
 	protected DeleteConflictService myDeleteConflictService;
 	@Autowired
@@ -1528,4 +1525,5 @@ public abstract class BaseHapiFhirDao<T extends IBaseResource> extends BaseStora
 				"Resource with ID " + theEntity.getIdDt().getIdPart() + " exists but it is not of type " + theResourceName + ", found resource of type " + theEntity.getResourceType());
 		}
 	}
+
 }
